requirements.txt
six==1.16.0
numpy==1.16.6
dbus-python==1.2.18
pygobject==3.36.1
requests==2.27.1
requests-oauthlib==1.3.1
flask==1.1.4
flask-bootstrap==3.3.7.1